refactor package naming and structure
