+++ src/main/kotlin/org/wycliffeassociates/sourceaudio/common/data/model/MediaQuality.kt
@@ -1,4 +1,4 @@
-package org.wycliffeassociates.sourceaudio.directorymaker.common.data.model
+package org.wycliffeassociates.sourceaudio.common.data.model
 
 enum class MediaQuality(val quality: String) {
     HI("hi"),
